updated to use or not use .pem for chrome if exists

+++ tools/make-artifacts.sh
@@ -4,6 +4,8 @@ CHROME=/Applications/Google\ Chrome.app/Contents/MacOS/Google\ Chrome
 FIREFOX=/Applications/FirefoxNightly.app/Contents/MacOS/firefox-bin
 OPERA=/Applications/Opera.app/Contents/MacOS/Opera
 
+CHROME_PEM=./adnauseam.chromium.pem
+
 # before running, check that above programs exist, as well as that 'webext' cmd exists
 # if not, exit with error
 command "${CHROME}" --version || { echo >&2 "Chrome is not installed."; exit 1; }
@@ -15,9 +17,7 @@ hash web-ext 2>/dev/null || { echo >&2 "Webext is not installed. Please do npm i
 DES=dist/build
 ARTS=artifacts
 
-CHROME_OPTS1=--pack-extension=${DES}/adnauseam.chromium
-CHROME_OPTS2=--pack-extension-key=adnauseam.chromium.pem
-
+CHROME_OPTS=--pack-extension=${DES}/adnauseam.chromium
 OPERA_OPTS=--pack-extension=${DES}/adnauseam.opera
 
 VERSION=`jq .version manifest.json | tr -d '"'`
@@ -31,8 +31,15 @@ rm -rf ${DES}/*
 
 # CHROME
 ./tools/make-chromium.sh
-"${CHROME}" "${CHROME_OPTS1}" "${CHROME_OPTS2}"
-mv ${DES}/adnauseam.chromium.crx ${ARTS}/adnauseam-${VERSION}.chromium.crx
+if [ -f $CHROME_PEM ]; then  # do we have the signing key?
+  "${CHROME}" "${CHROME_OPTS}" "--pack-extension-key $CHROME_PEM"
+  mv ${DES}/adnauseam.chromium.crx ${ARTS}/adnauseam-${VERSION}.chromium.crx
+  echo "*** AdNauseam::Chromium: Signed with local .pem\\n"
+else
+  "${CHROME}" "${CHROME_OPTS}"
+  mv ${DES}/adnauseam.chromium.crx ${ARTS}/adnauseam-${VERSION}.chromium-UNSIGNED.crx
+  echo "WARN: NO .pem key found for Chrome build, this is required for publishing\\n"
+fi
 
 
 # OPERA
@@ -52,12 +59,10 @@ cp ${DES}/adnauseam.firefox/null.xpi ${ARTS}/adnauseam-${VERSION}.firefox-legacy
 # WEBEXT
 ./tools/make-webext.sh all
 web-ext build -s ${DES}/adnauseam.webext -a ${ARTS}
-mv ${ARTS}/adnauseam-${VERSION}.zip ${ARTS}/adnauseam-${VERSION}.firefox-webext.zip
+mv ${ARTS}/adnauseam-${VERSION}.zip ${ARTS}/adnauseam-${VERSION}.firefox.zip
 
 # NO PEMS
 mv ${DES}/*.pem /tmp
 
 ls -l artifacts
 open artifacts
-
-#echo Pack/sign, then do $ cp dist/build/adnauseam.chromium.crx artifacts/adnauseam-${VERSION}.chromium.crx
